feat: Add Matic token contract
Add a new contract called Matic.sol which implements the Matic token functionality. This contract extends ERC20, ERC20Burnable, Ownable, and ERC20Permit from the OpenZeppelin Contracts library. It includes a mapping to track token balances of holders and a mint function that can only be called by the contract owner. The contract is named "Matic" with the symbol "MATIC". The contract also includes commented out code for a custom transfer function that updates the balances mapping.

SPDX-License-Identifier: MIT
Compatible with OpenZeppelin Contracts ^5.0.0